Refactor prover crate
