+++ .github/workflows/pr-main_l2.yaml
@@ -79,7 +79,7 @@ jobs:
         run: |
           cd crates/l2
           make build-prover
-          mkdir -p prover/zkvm/interface/sp1/out && touch prover/zkvm/interface/sp1/out/riscv32im-succinct-zkvm-vk
+          mkdir -p prover/src/guest_program/src/sp1/out && touch prover/src/guest_program/src/sp1/out/riscv32im-succinct-zkvm-vk
 
       - name: Build test
         run: |
@@ -156,7 +156,7 @@ jobs:
         run: |
           cd crates/l2
           make build-prover
-          mkdir -p prover/zkvm/interface/sp1/out && touch prover/zkvm/interface/sp1/out/riscv32im-succinct-zkvm-vk
+          mkdir -p prover/src/guest_program/src/sp1/out && touch prover/src/guest_program/src/sp1/out/riscv32im-succinct-zkvm-vk
 
       - name: Start L1 & Deploy contracts
         run: |
