+++ .github/workflows/tag_release.yaml
@@ -139,7 +139,7 @@ jobs:
             --bin ethrex_prover
           cd ../../
           mv target/release/ethrex_prover ethrex_prover_sp1-${{ matrix.os }}_${{ matrix.arch }}
-          mv crates/l2/prover/zkvm/interface/sp1/out/riscv32im-succinct-zkvm-vk ethrex-riscv32im-succinct-zkvm-vk
+          mv crates/l2/prover/src/guest_program/src/sp1/out/riscv32im-succinct-zkvm-vk ethrex-riscv32im-succinct-zkvm-vk
 
       - name: Upload artifact
         uses: actions/upload-artifact@v4
@@ -190,7 +190,7 @@ jobs:
             --bin ethrex_prover
           cd ../../
           mv target/release/ethrex_prover ethrex_prover_risc0-linux_x86_64
-          mv crates/l2/prover/zkvm/interface/risc0/out/riscv32im-risc0-vk ethrex-riscv32im-risc0-vk
+          mv crates/l2/prover/src/guest_program/src/risc0/out/riscv32im-risc0-vk ethrex-riscv32im-risc0-vk
 
       - name: Upload artifact
         uses: actions/upload-artifact@v4
