+++ cmd/ef_tests/blockchain/Cargo.toml
@@ -10,7 +10,7 @@ ethrex-storage.workspace = true
 ethrex-rlp.workspace = true
 ethrex-vm.workspace = true
 ethrex-prover = { path = "../../../crates/l2/prover", default-features = false }
-zkvm_interface = { path = "../../../crates/l2/prover/zkvm/interface", default-features = false }
+zkvm_interface = { path = "../../../crates/l2/prover/src/guest_program", default-features = false }
 
 serde.workspace = true
 serde_json.workspace = true
