+++ cmd/ethrex_replay/Cargo.toml
@@ -10,7 +10,7 @@ ethrex-trie.workspace = true
 ethrex-levm.workspace = true
 ethrex-rpc.workspace = true
 
-zkvm_interface = { path = "../../crates/l2/prover/zkvm/interface", default-features = false }
+zkvm_interface = { path = "../../crates/l2/prover/src/guest_program", default-features = false }
 ethrex-prover = { path = "../../crates/l2/prover", default-features = false }
 
 serde.workspace = true
