+++ crates/l2/Cargo.toml
@@ -55,7 +55,7 @@ tui-logger.workspace = true
 tabwriter = "1.4.1"
 color-eyre = "0.6.5"
 
-zkvm_interface = { path = "./prover/zkvm/interface/" }
+zkvm_interface = { path = "./prover/src/guest_program/" }
 
 
 [build-dependencies]
