+++ crates/l2/docker-compose-l2.yaml
@@ -18,8 +18,8 @@ services:
       - ../../fixtures/genesis/l1-dev.json:${CI_ETHREX_WORKDIR}/fixtures/genesis/l1-dev.json
       - ../../fixtures/genesis/l2.json:${CI_ETHREX_WORKDIR}/fixtures/genesis/l2.json
       - ../../fixtures/keys/private_keys_l1.txt:${CI_ETHREX_WORKDIR}/fixtures/keys/private_keys_l1.txt
-      - ./prover/zkvm/interface/sp1/out/riscv32im-succinct-zkvm-vk:${CI_ETHREX_WORKDIR}/riscv32im-succinct-zkvm-vk
-      - ./prover/zkvm/interface/risc0/out/riscv32im-risc0-vk:${CI_ETHREX_WORKDIR}/riscv32im-risc0-vk
+      - ./prover/src/guest_program/src/sp1/out/riscv32im-succinct-zkvm-vk:${CI_ETHREX_WORKDIR}/riscv32im-succinct-zkvm-vk
+      - ./prover/src/guest_program/src/risc0/out/riscv32im-risc0-vk:${CI_ETHREX_WORKDIR}/riscv32im-risc0-vk
     environment:
       - ETHREX_ETH_RPC_URL=http://ethrex_l1:8545
       # NOTE: The paths in the env variables must match those
