--- crates/l2/prover/.gitignore
@@ -1,2 +1,0 @@
-rsp/
-target/
