+++ crates/l2/prover/Cargo.toml
@@ -32,7 +32,7 @@ ethrex-l2 = { path = "../", default-features = false }
 ethrex-l2-common = { workspace = true }
 ethrex-sdk.workspace = true
 
-zkvm_interface = { path = "./zkvm/interface", default-features = false }
+zkvm_interface = { path = "./src/guest_program", default-features = false }
 
 risc0-zkvm = { version = "2.1.0", optional = true }
 risc0-zkp = {version = "2.0.1", optional = true }
