+++ crates/l2/prover/src/backends/sp1.rs
@@ -1,18 +1,18 @@
-use sp1_sdk::{
-    EnvProver, HashableKey, ProverClient, SP1ProofWithPublicValues, SP1ProvingKey, SP1Stdin,
-    SP1VerifyingKey,
-};
 use std::{fmt::Debug, sync::LazyLock};
-use tracing::info;
-use zkvm_interface::io::{JSONProgramInput, ProgramInput};
 
 use ethrex_l2_common::{
     calldata::Value,
     prover::{BatchProof, ProofBytes, ProofCalldata, ProverType},
 };
+use sp1_sdk::{
+    EnvProver, HashableKey, ProverClient, SP1ProofWithPublicValues, SP1ProvingKey, SP1Stdin,
+    SP1VerifyingKey,
+};
+use tracing::info;
+use zkvm_interface::input::{JSONProgramInput, ProgramInput};
 
 static PROGRAM_ELF: &[u8] =
-    include_bytes!("../../zkvm/interface/sp1/out/riscv32im-succinct-zkvm-elf");
+    include_bytes!("../guest_program/src/sp1/out/riscv32im-succinct-zkvm-elf");
 
 struct ProverSetup {
     client: EnvProver,
