+++ crates/l2/prover/src/guest_program/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod execution;
-pub mod io;
+pub mod input;
 pub mod methods;
+pub mod output;
