+++ crates/l2/prover/src/guest_program/src/sp1/NOTICE
@@ -1,4 +1,5 @@
 # APACHE NOTICE
+
 Original work Copyright [Succinct Labs]  
 Copyright [2024] [LambdaClass]
 
@@ -14,8 +15,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 
-
 # MIT NOTICE
+
 The MIT License (MIT)
 
 Original work Copyright (c) 2023 Succinct Labs
