--- crates/l2/prover/zkvm/.gitignore
@@ -1,4 +1,0 @@
-.DS_Store
-target/
-interface/sp1/out
-interface/risc0/out
