--- crates/l2/prover/zkvm/interface/risc0/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "zkvm-risc0-program"
-version = "0.1.0"
-edition = "2024"
-
-[workspace]
-
-[dependencies]
-risc0-zkvm = { version = "2.1.0", default-features = false, features = ["std"] }
-risc0-zkvm-platform = { version = "2.0.2", default-features = false, features = ["sys-getenv"] }
-zkvm_interface = { path = "../", default-features = false, features = ["c-kzg"]}
-
-ethrex-common = { path = "../../../../../common", default-features = false }
-ethrex-storage = { path = "../../../../../storage", default-features = false }
-ethrex-rlp = { path = "../../../../../common/rlp" }
-ethrex-vm = { path = "../../../../../vm", default-features = false, features = ["c-kzg"]}
-ethrex-blockchain = { path = "../../../../../blockchain", default-features = false }
-ethrex-l2-common = { path = "../../../../common", default-features = false }
-
-[patch.crates-io]
-sha2 = { git = "https://github.com/risc0/RustCrypto-hashes", tag = "sha2-v0.10.8-risczero.0" }
-c-kzg = { git = "https://github.com/risc0/c-kzg-4844", tag = "c-kzg/v1.0.3-risczero.1" }
-# This git repo seems to be private
-# ed25519-dalek = { git = "https://github.com/risc0/ed25519-dalek", tag = "curve25519-4.1.2-risczero.0" }
-
-# Other precompiles/patches can be added, but they require the "unstable" risc0 feature which is not suited
-# for production environments.
-
-[features]
-l2 = ["zkvm_interface/l2"]
