+++ crates/l2/tee/quote-gen/Cargo.toml
@@ -18,7 +18,7 @@ ethrex-rlp = { path = "../../../common/rlp" }
 ethrex-vm = { path = "../../../vm", default-features = false }
 ethrex-blockchain = { path = "../../../blockchain", default-features = false }
 ethrex-rpc = { path = "../../../networking/rpc", default-features = false }
-zkvm_interface = { path = "../../prover/zkvm/interface", default-features = false }
+zkvm_interface = { path = "../../prover/src/guest_program", default-features = false }
 ethrex-sdk = { path = "../../sdk", default-features = false }
 ethrex-l2 = { path = "../..", default-features = false }
 ethrex-l2-common = { path = "../../common", default-features = false }
