+++ docs/l2/prover.md
@@ -5,7 +5,7 @@
 
 ## Intro
 
-The prover consists of two main components: handling incoming proving data from the `L2 proposer`, specifically from the `ProofCoordinator` component, and the `zkVM`. The `Prover` is responsible for this first part, while the `zkVM` serves as a RISC-V emulator executing code specified in `crates/l2/prover/zkvm/interface/guest/src`.
+The prover consists of two main components: handling incoming proving data from the `L2 proposer`, specifically from the `ProofCoordinator` component, and the `zkVM`. The `Prover` is responsible for this first part, while the `zkVM` serves as a RISC-V emulator executing code specified in `crates/l2/prover/src/guest_program/guest/src`.
 Before the `zkVM` code (or guest), there is a directory called `interface`, which indicates that we access the `zkVM` through the "interface" crate.
 
 In summary, the `Prover` manages the inputs from the `ProofCoordinator` and then "calls" the `zkVM` to perform the proving process and generate the `groth16` ZK proof.
@@ -31,7 +31,7 @@ sequenceDiagram
 
 **Dependencies:**
 
-- [RISC0](https://dev.risczero.com/api/zkvm/install)
+- [RISC0](https://dev.risczero.com/api/src/guest_program/install)
   1. `curl -L https://risczero.com/install | bash`
   2. `rzup install cargo-risczero 1.2.0`
 - [SP1](https://docs.succinct.xyz/docs/sp1/introduction)
@@ -78,7 +78,7 @@ After this initialization we should have the prover running in `dev_mode` &rarr;
    - Add the `user` you are using to the `docker` group &rarr; command: `sudo usermod -aG docker $USER`. (needs reboot, doing it after CUDA installation)
    - `id -nG` after reboot to check if the user is in the group.
 2. Install [Rust](https://www.rust-lang.org/tools/install)
-3. Install [RISC0](https://dev.risczero.com/api/zkvm/install)
+3. Install [RISC0](https://dev.risczero.com/api/src/guest_program/install)
 4. Install [CUDA for Ubuntu](https://developer.nvidia.com/cuda-downloads?target_os=Linux&target_arch=x86_64&Distribution=Ubuntu&target_version=22.04&target_type=deb_local)
    - Install `CUDA Toolkit Installer` first. Then the `nvidia-open` drivers.
 5. Reboot
